chore cleanup CarouselContext
