+++ packages/react-components/react-teaching-popover-preview/src/components/TeachingPopoverCarousel/Carousel/createCarouselStore.ts
@@ -1,11 +1,5 @@
-import { Context, ContextSelector, createContext, useContextSelector } from '@fluentui/react-context-selector';
-import { CarouselPageChangeData, CarouselStore } from './Carousel.types';
-import * as React from 'react';
-import { EventHandler } from '@fluentui/react-utilities';
+import { type CarouselStore } from './Carousel.types';
 
-/**
- * A hook for managing a collection of carousel pages
- */
 export const createCarouselStore = (): CarouselStore => {
   let values: string[] = [];
   let listeners: Array<() => void> = [];
@@ -59,38 +53,5 @@ export const createCarouselStore = (): CarouselStore => {
     }
   }
 
-  return {
-    ...carouselStore,
-  };
+  return carouselStore;
 };
-
-export const useCarouselCollection_unstable = () => {
-  const [store] = React.useState(() => createCarouselStore());
-
-  return {
-    store,
-  };
-};
-
-export type CarouselContextValue = {
-  store: CarouselStore;
-  value: string;
-  setValue: (value: string) => void;
-  onPageChange?: EventHandler<CarouselPageChangeData>;
-};
-
-export const carouselContextDefaultValue: CarouselContextValue = {
-  store: createCarouselStore(),
-  value: '',
-  setValue: () => {},
-  onPageChange: () => null,
-};
-
-export const CarouselContext: Context<CarouselContextValue> = createContext<CarouselContextValue | undefined>(
-  undefined,
-) as Context<CarouselContextValue>;
-
-export const CarouselProvider = CarouselContext.Provider;
-
-export const useCarouselContext_unstable = <T>(selector: ContextSelector<CarouselContextValue, T>): T =>
-  useContextSelector(CarouselContext, (ctx = carouselContextDefaultValue) => selector(ctx));
